refactor: 무한스크롤 컴포넌트 구현 및 로딩 컴포넌트 삭제
