+++ src/components/molecules/index.ts
@@ -1,5 +1,5 @@
 export { ContentWrapper } from './contentWrapper';
-export { LoadingInfiniteScroll } from './infiniteScroll';
+export { InfiniteScroller } from './infiniteScroll';
 export { LimitedLengthInput } from './limitedLengthInput';
 export { LoginIconSet } from './loginIconSet';
 export { SelectableCardList } from './selectableCardList';
